refactor: moved useMemberSummary function into its own file
