+++ src/providers/MemberData.jsx
@@ -207,44 +207,6 @@ export function useMemberData(address, chainId) {
   };
 }
 
-export function useMemberSummary(address, chainId) {
-  const userManagerContract = useContract("userManager", chainId);
-
-  const contracts = address
-    ? [
-      {
-        ...userManagerContract,
-        functionName: "checkIsMember",
-        args: [address],
-      },
-      {
-        ...userManagerContract,
-        functionName: "getCreditLimit",
-        args: [address],
-      },
-    ]
-    : [];
-
-  const { data, ...resp } = useContractReads({
-    enabled: !!address,
-    select: (data) => ({
-      isMember: data[0] || false,
-      creditLimit: data[1] || ZERO,
-    }),
-    contracts: contracts.map((contract) => ({
-      ...contract,
-      chainId,
-    })),
-    cacheTime: CACHE_TIME,
-    staleTime: Infinity,
-  });
-
-  return {
-    data: { ...data, address },
-    ...resp,
-  };
-}
-
 export default function MemberData({
   chainId,
   children,
